crondb2/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
greenlet==3.0.3
    # via sqlalchemy
ibm-db==3.2.1
    # via
    #   -r requirements.in
    #   ibm-db-sa
ibm-db-sa==0.4.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
schedule==1.2.1
    # via -r requirements.in
sqlalchemy==2.0.24
    # via ibm-db-sa
typing-extensions==4.9.0
    # via sqlalchemy
